+++ src/PSCompression/Algorithm.cs
@@ -3,7 +3,6 @@ namespace PSCompression;
 public enum Algorithm
 {
     gz,
-    br,
     bz2,
     zst,
     lz,